Move stuff around in settings, add new specating status, add spectate button in profile, add tooltip with status name in profile
