+++ frontend/src/app.css
@@ -1,3 +1,11 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
+
+.custom-tooltip {
+  @apply invisible absolute;
+}
+
+.has-tooltip:hover .custom-tooltip {
+  @apply visible z-50;
+}
