+++ frontend/src/lib/ProfilePic.svelte
@@ -16,21 +16,26 @@
       .catch(() => `https://cdn.intra.42.fr/users/${u}.jpg`);
 
   // typescript hack because it can't find the type of pfp and img src needs to be a string
-  let img: string;
   $: img = $pfp as string;
 
-  const fn = cond<[number], string>([
-    [equals(1), always("badge-success")],
-    [equals(2), always("badge-error")],
-    [T, always("")],
+  const getStatus = cond<[number], [string, string]>([
+    [equals(1), always(["badge-success", "Online"])],
+    [equals(2), always(["badge-error", "Playing"])],
+    [equals(3), always(["badge-primary", "Spectating"])],
+    [T, always(["", "Offline"])],
   ]);
 
-  const colour = fn(status);
+  const [colour, text] = getStatus(status);
 </script>
 
 <div class={status !== -1 ? "indicator" : ""}>
   {#if status !== -1}
-    <span class="indicator-item indicator-bottom badge {colour}" />
+    <span class="has-tooltip indicator-item indicator-bottom badge {colour}">
+      <span
+        class="custom-tooltip rounded shadow-lg bg-base-300 p-2 mt-16 text-white"
+        >{text}</span
+      >
+    </span>
   {/if}
   {#if user === $login}
     <img src={img} class={attributes} alt="userPFP" />
